Feature/generate report as excel

+++ src/etl/tasks/download_document.py
@@ -1,6 +1,9 @@
 """This script contain prefect task to download the asked document
 """
+import os
+import datetime
 from prefect import task
+from prefect.engine.signals import SKIP
 import urllib.request
 from .consts import DOCUMENT_URL, DOWNLOAD_PATH
 
@@ -16,10 +19,16 @@ HEADERS = {
 }
 
 
-@task(log_stdout=True)
+@task(
+    log_stdout=True, max_retries=3, retry_delay=datetime.timedelta(minutes=2),
+)
 def download_document():
-    print(f"Downloading the file: {DOCUMENT_URL}, to: {DOWNLOAD_PATH}")
-    opener = urllib.request.build_opener()
-    opener.addheaders = HEADERS.items()
-    urllib.request.install_opener(opener)
-    urllib.request.urlretrieve(DOCUMENT_URL, DOWNLOAD_PATH)
+    if not os.path.exists(DOWNLOAD_PATH):
+        print(f"Downloading the file: {DOCUMENT_URL}, to: {DOWNLOAD_PATH}")
+        opener = urllib.request.build_opener()
+        opener.addheaders = HEADERS.items()
+        urllib.request.install_opener(opener)
+        urllib.request.urlretrieve(DOCUMENT_URL, DOWNLOAD_PATH)
+    else:
+        print(f"File: {DOWNLOAD_PATH} already exists, skipping this task")
+        raise SKIP
